+++ go.mod
@@ -5,7 +5,7 @@ go 1.17
 require (
 	github.com/armosec/armoapi-go v0.0.23
 	github.com/armosec/k8s-interface v0.0.32
-	github.com/armosec/opa-utils v0.0.61
+	github.com/armosec/opa-utils v0.0.62
 	github.com/armosec/rbac-utils v0.0.8
 	github.com/armosec/utils-go v0.0.3
 	github.com/briandowns/spinner v1.16.0
